refactor(app): split flows, narrow DB protocol, remove Rust
Non-embedding cleanup carried over from source commit 44e2e83:

- Split oversized flows: extract the LLM retrieve pipeline into retrieve_llm.py
  (RetrieveLlmMixin) and the memorize response parsers into memorize_parse.py
  (MemorizeParseMixin); compose both onto MemoryService. memorize.py drops the
  parser bodies and their now-unused json/xml imports.
- Narrow the Database protocol to repositories only (no raw record dicts) so
  scope filtering cannot be bypassed.
- Remove the unused Rust scaffolding (lib.rs/Cargo/maturin) and switch the build
  backend to hatchling.

The embedding call sites in memorize.py are deliberately left on the existing
(_get_step_embedding_client / chat-client embed) API; the embedding decouple is
isolated in the final step. memorize() remains byte-identical to main.
