--- Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "memu"
-version = "0.1.0"
-edition = "2024"
-
-[lib]
-name = "_core"
-# "cdylib" is necessary to produce a shared library for Python to import from.
-crate-type = ["cdylib"]
-
-[dependencies]
-# "extension-module" tells pyo3 we want to build an extension module (skips linking against libpython.so)
-# "abi3-py313" tells pyo3 (and maturin) to build using the stable ABI with minimum Python version 3.13
-pyo3 = { version = "0.27.1", features = ["extension-module", "abi3-py313"] }
