+++ src/memu/database/postgres/postgres.py
@@ -91,11 +91,6 @@ class PostgresStore(Database):
             scope_fields=self._scope_fields,
         )
 
-        self.resources = self._state.resources
-        self.items = self._state.items
-        self.categories = self._state.categories
-        self.relations = self._state.relations
-
         # self._load_existing()
 
     def close(self) -> None:
